Cleaned up OPEMetadata and PPKey
